refactor: split digital-twin routes into domain sub-router directory (#2533)
Break the ~990-line flat server/routes/digital-twin.js multi-domain router
into a digital-twin/ directory of thirteen domain sub-routers assembled by
index.js, mirroring the apps/pipeline sub-router pattern. The shared
assertPersonaExists guard moves to shared.js. All 73 endpoints keep their
exact method+path, Zod validation, and ServerError status mapping; the route
test moves into the directory as index.test.js.

Closes #2533
